+++ src/ppu/internal.h
@@ -3,6 +3,6 @@
 #include <stdint.h>
 
 // Writes to pixel buffer
-void w_pxl(int y, int x, int clr);
+void w_pxl(int y, int x, uint8_t clr);
 
-int gt_clr(uint8_t pal, int val);
+uint8_t gt_clr(uint8_t pal, int val);
